Moves page library
